v3 - intermediate non working.
for some reason, I cant get the process to close stdout and stderr
before ending. It has something to do with the async-ness of Deno.run
I wish it could just offer a synchronous mode. For "paperback" we need
each step to complete before moving on to the next, because each depends
upon the previous.